ModifyLookupAsyncShouldSetAuditFieldsCorrectly -> WIP
