+++ ISL.ReIdentification.Core.Tests.Unit/Services/Foundations/Lookups/LookupServiceTests.Logic.ApplyModifyAudit.cs
@@ -1,0 +1,61 @@
+// ---------------------------------------------------------
+// Copyright (c) North East London ICB. All rights reserved.
+// ---------------------------------------------------------
+
+using System;
+using System.Threading.Tasks;
+using FluentAssertions;
+using ISL.ReIdentification.Core.Models.Foundations.Lookups;
+using ISL.ReIdentification.Core.Models.Securities;
+using ISL.ReIdentification.Core.Services.Foundations.Lookups;
+using Moq;
+
+namespace ISL.ReIdentification.Core.Tests.Unit.Services.Foundations.Lookups
+{
+    public partial class LookupServiceTests
+    {
+        [Fact]
+        public async Task DeleteLookupAsyncShouldSetAuditFieldsCorrectly()
+        {
+            // Given
+            DateTimeOffset randomDateTimeOffset = GetRandomDateTimeOffset();
+            DateTimeOffset randomLookupDate = GetRandomDateTimeOffset();
+            EntraUser randomEntraUser = CreateRandomEntraUser();
+            Lookup randomLookup = CreateRandomLookup(randomLookupDate, GetRandomString());
+
+            this.dateTimeBrokerMock.Setup(broker =>
+                broker.GetCurrentDateTimeOffsetAsync())
+                    .ReturnsAsync(randomDateTimeOffset);
+
+            this.securityBrokerMock.Setup(broker =>
+                broker.GetCurrentUserAsync())
+                    .ReturnsAsync(randomEntraUser);
+
+            var lookupServiceMock = new Mock<LookupService>(
+                this.reIdentificationStorageBroker.Object,
+                this.dateTimeBrokerMock.Object,
+                this.securityBrokerMock.Object,
+                this.loggingBrokerMock.Object);
+
+            // When
+            Lookup actualLookup = await lookupServiceMock.Object.ApplyModifyAuditAsync(randomLookup);
+
+            // Then
+            actualLookup.UpdatedBy.Should().BeEquivalentTo(randomEntraUser.EntraUserId.ToString());
+            actualLookup.UpdatedDate.Should().Be(randomDateTimeOffset);
+
+            this.dateTimeBrokerMock.Verify(broker =>
+                broker.GetCurrentDateTimeOffsetAsync(),
+                    Times.Once);
+
+            this.securityBrokerMock.Verify(broker =>
+                broker.GetCurrentUserAsync(),
+                    Times.Once);
+
+            this.dateTimeBrokerMock.VerifyNoOtherCalls();
+            this.securityBrokerMock.VerifyNoOtherCalls();
+            this.loggingBrokerMock.VerifyNoOtherCalls();
+            this.reIdentificationStorageBroker.VerifyNoOtherCalls();
+        }
+    }
+}
